+++ examples/llm-example/mcp-gateway/docker-compose.yml
