+++ examples/llm-example/src/main/kotlin/com/justai/jaicf/examples/llm/AgentWithMcpServiceSse.kt
@@ -1,10 +1,7 @@
 package com.justai.jaicf.examples.llm
 
-import com.justai.jaicf.BotEngine
+import com.justai.jaicf.activator.llm.agent.LLMAgent
 import com.justai.jaicf.activator.llm.mcp.McpService
-import com.justai.jaicf.activator.llm.mcp.getTool
-import com.justai.jaicf.activator.llm.scenario.llmState
-import com.justai.jaicf.builder.Scenario
 import com.justai.jaicf.examples.llm.channel.ConsoleChannel
 
 /**
@@ -17,7 +14,7 @@ import com.justai.jaicf.examples.llm.channel.ConsoleChannel
  * - **Streaming** (`--transport=streaming`) - HTTP with streaming support, use with `McpService.streamableHttp()`
  *
  * ## Setup:
- * 1. See docker-compose.yml at: activators/llm/src/main/kotlin/com/justai/jaicf/activator/llm/mcp/gateway/docker-compose.yml
+ * 1. See docker-compose.yml at: activators/llm/mcp-gateway/docker-compose.yml
  * 2. Start the gateway: `docker-compose up`
  * 3. Connect to the gateway using one of the transports below
  *
@@ -28,11 +25,11 @@ import com.justai.jaicf.examples.llm.channel.ConsoleChannel
  *   EXA_API_KEY: your-api-key-here
  * ```
  *
- * For STDIO transport examples, see ScenarioWithMcpServiceStdio.kt
+ * For STDIO transport examples, see AgentWithMcpServiceStdio.kt
  */
 
-val mcpServiceSse = McpService.sse(
-    urlString = "http://localhost:8081/sse"
+val searchMCPService = McpService.sse(
+    url = "http://localhost:8081/sse"
 )
 
 /**
@@ -41,12 +38,13 @@ val mcpServiceSse = McpService.sse(
  *
  * IMPORTANT! Set up your OPENAI_API_KEY and OPENAI_BASE_URL env before running
  */
-private val scenario = Scenario {
-    llmState("main", {
+private val agent = LLMAgent(
+    name = "mcp-agent",
+    props = {
         model = "gpt-4.1-nano"
 
         // The most simple way to add a tool. Provide a tool name via `toolName` parameter.
-        tool(mcpServiceSse.getTool("search"))
+        tool(searchMCPService.tool("search"))
 
         /**
          * You can also use other options for your scenario:
@@ -84,10 +82,8 @@ private val scenario = Scenario {
          * }
          * ```
          */
-    })
-}
-
+    }
+)
 fun main() {
-    ConsoleChannel(BotEngine(scenario))
-        .run("Hi")
+    ConsoleChannel(agent.asBot).run("What tools you have?")
 }
